Refactor: Organize Navigation Code and Add Breadcrumb Logging
This commit refactors the navigation-related code and introduces breadcrumb logging for Firebase.

- **Navigation Code Reorganization:**
    - `NavGraph.kt` has been moved from `kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation` to `kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/navigation`. The package declaration has been updated accordingly.

- **Firebase Breadcrumb Logging:**
    - **`kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/navigation/NavigationUtils.kt`**:
        - Introduced a new composable function `AddBreadcrumbLogging(navController: NavController)`.
        - This function uses `DisposableEffect` to add an `OnDestinationChangedListener` to the `navController`.
        - When the destination changes, it logs a Firebase message with the route of the new destination (e.g., "Navigated to: [destination.route]").
        - This logging is only active if `appConfig.buildType` is not `BuildType.NoFirebase`.
        - The listener is removed when the composable leaves the composition.
    - **`UIViews/src/main/java/com/programmersbox/uiviews/presentation/navigation/NavGraphView.kt`**:
        - Called `AddBreadcrumbLogging(navController)` within the `NavHost` setup, enabling breadcrumb logging for the Android app.

- **Desktop UI Navigation Update:**
    - **`kmpuiviews/src/jvmMain/kotlin/com/programmersbox/kmpuiviews/DesktopUi.kt`**:
        - Replaced the direct `NavHost` usage with a new `HomeNav` composable (presumably defined elsewhere) which now wraps the `NavHost` and its `navGraph` configuration. This change likely centralizes or standardizes navigation setup for the desktop target.
        - The `navGraph` call remains largely the same but is now nested within `HomeNav`.
        - Imports for `com.programmersbox.kmpuiviews.presentation.HomeNav` and updated `navGraph` path were added.

- **Minor Import Update:**
    - **`kmpuiviews/src/androidMain/kotlin/com/programmersbox/kmpuiviews/Platform.android.kt`**:
        - Added an import for `androidx.navigation.NavController`.
